// frontend/src/lib/types.ts
export type UserRole = "admin" | "marketing_manager" | "telecaller" | "counselor" | "telecalling_manager" | "owner" | "alliance_manager" | "alliance_executive" | "accounts_manager" | "accounts_executive";

export interface User {
  id: string;
  name: string;
  email: string;
  role: UserRole;
}

export interface AuthLoginResponse {
  _id: string;
  name: string;
  email: string;
  role: UserRole;
  token: string;
}